testRefactor : Section 생성 시 LineClient 를 사용하도록 리팩토링
