+++ src/test/java/nextstep/subway/LineClient.java
@@ -47,4 +47,18 @@ public class LineClient {
                 .then().log().all().extract();
     }
 
+    public void addSection(String downStationId, String upStationId, int distance, long id) {
+        final var params = Map.of(
+                "downStationId", downStationId,
+                "upStationId", upStationId,
+                "distance", distance
+        );
+        RestAssured.given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/lines/" + id + "/sections")
+                .then().log().all()
+                .extract();
+    }
+
 }
